+++ firebase-vscode/src/data-connect/emulator.ts
@@ -1,26 +1,47 @@
 import { EmulatorsController } from "../core/emulators";
 import * as vscode from "vscode";
 import { ExtensionBrokerImpl } from "../extension-broker";
-import { Signal, effect, signal } from "@preact/signals-core";
-import { RC } from "../rc";
+import { effect, signal } from "@preact/signals-core";
+import { firstWhereDefined } from "../utils/signal";
 import { firebaseRC, updateFirebaseRCProject } from "../core/config";
+import {
+  DataConnectEmulatorClient,
+  dataConnectEmulatorEvents,
+} from "../../../src/emulator/dataconnectEmulator";
+import { dataConnectConfigs } from "./config";
+import { runEmulatorIssuesStream } from "./emulator-stream";
+import { runDataConnectCompiler } from "./core-compiler";
+import { pluginLogger } from "../logger-wrapper";
+import { Emulators, getEmulatorDetails, listRunningEmulators } from "../cli";
+import { emulatorOutputChannel } from "./emulator-stream";
 
 /** FDC-specific emulator logic */
 export class DataConnectEmulatorController implements vscode.Disposable {
   constructor(
     readonly emulatorsController: EmulatorsController,
-    broker: ExtensionBrokerImpl,
+    readonly broker: ExtensionBrokerImpl,
   ) {
     function notifyIsConnectedToPostgres(isConnected: boolean) {
       broker.send("notifyIsConnectedToPostgres", isConnected);
     }
 
+    // on emulator restart, re-connect
+    dataConnectEmulatorEvents.on("restart", () => {
+      pluginLogger.log("Emulator started");
+      this.isPostgresEnabled.value = false;
+      this.connectToEmulator();
+    });
+
     this.subs.push(
       broker.on("connectToPostgres", () => this.connectToPostgres()),
-      broker.on("disconnectPostgres", () => this.disconnectPostgres()),
 
       // Notify webviews when the emulator status changes
       effect(() => {
+        if (this.isPostgresEnabled.value) {
+          this.emulatorsController.emulatorStatusItem.show();
+        } else {
+          this.emulatorsController.emulatorStatusItem.hide();
+        }
         notifyIsConnectedToPostgres(this.isPostgresEnabled.value);
       }),
 
@@ -34,12 +55,14 @@ export class DataConnectEmulatorController implements vscode.Disposable {
   readonly isPostgresEnabled = signal(false);
   private readonly subs: Array<() => void> = [];
 
-  private async promptConnectionString(): Promise<string | undefined> {
+  private async promptConnectionString(
+    defaultConnectionString: string,
+  ): Promise<string | undefined> {
     const connectionString = await vscode.window.showInputBox({
       title: "Enter a Postgres connection string",
       prompt:
-        "A Postgres database must be configured to use the emulator locally. ",
-      placeHolder: "postgres://user:password@localhost:5432/dbname",
+        "A Postgres database must be configured to use the emulator locally.",
+      value: defaultConnectionString,
     });
 
     return connectionString;
@@ -47,25 +70,96 @@ export class DataConnectEmulatorController implements vscode.Disposable {
 
   private async connectToPostgres() {
     const rc = firebaseRC.value?.tryReadValue;
-
-    if (!rc?.getDataconnect()?.postgres) {
-      const newConnectionString = await this.promptConnectionString();
-      if (!newConnectionString) {
-        return;
+    let localConnectionString =
+      rc?.getDataconnect()?.postgres?.localConnectionString;
+    if (!localConnectionString) {
+      const dataConnectConfigsValue =
+        await firstWhereDefined(dataConnectConfigs);
+      let dbname = "postgres";
+      const postgresql =
+        dataConnectConfigsValue?.tryReadValue.values[0]?.value?.schema
+          ?.datasource?.postgresql;
+      if (postgresql) {
+        const instanceId = postgresql.cloudSql?.instanceId;
+        const databaseName = postgresql.database;
+        if (instanceId && databaseName) {
+          dbname = `${instanceId}-${databaseName}`;
+        }
       }
-
-      updateFirebaseRCProject({
-        fdcPostgresConnectionString: newConnectionString,
-      });
+      localConnectionString = `postgres://user:password@localhost:5432/${dbname}`;
+    }
+    const newConnectionString = await this.promptConnectionString(
+      localConnectionString,
+    );
+    if (!newConnectionString) {
+      return;
     }
 
+    // notify sidebar webview of connection string
+    this.broker.send("notifyPostgresStringChanged", newConnectionString);
+
+    updateFirebaseRCProject({
+      fdcPostgresConnectionString: newConnectionString,
+    });
+
+    // configure the emulator to use the local psql string
+    const emulatorClient = new DataConnectEmulatorClient();
     this.isPostgresEnabled.value = true;
-    this.emulatorsController.emulatorStatusItem.show();
+
+    emulatorClient.configureEmulator({ connectionString: newConnectionString });
   }
 
-  private disconnectPostgres() {
-    this.isPostgresEnabled.value = false;
-    this.emulatorsController.emulatorStatusItem.hide();
+  // on schema reload, restart language server and run introspection again
+  private async schemaReload() {
+    vscode.commands.executeCommand("fdc-graphql.restart");
+    vscode.commands.executeCommand(
+      "firebase.dataConnect.executeIntrospection",
+    );
+  }
+
+  // Commands to run after the emulator is started successfully
+  private async connectToEmulator() {
+    this.schemaReload();
+    const configs = dataConnectConfigs.value?.tryReadValue;
+
+    runEmulatorIssuesStream(
+      configs,
+      this.emulatorsController.getLocalEndpoint().value,
+      this.isPostgresEnabled,
+    );
+    runDataConnectCompiler(
+      configs,
+      this.emulatorsController.getLocalEndpoint().value,
+    );
+
+    this.setupOutputChannel();
+  }
+
+  private setupOutputChannel() {
+    if (
+      listRunningEmulators().filter((emulatorInfos) => {
+        emulatorInfos.name === Emulators.DATACONNECT;
+      })
+    ) {
+      const dataConnectEmulatorDetails = getEmulatorDetails(
+        Emulators.DATACONNECT,
+      );
+
+      // Child process is only available when emulator is started through vscode
+      if (dataConnectEmulatorDetails.instance) {
+        dataConnectEmulatorDetails.instance.stdout?.on("data", (data) => {
+          emulatorOutputChannel.appendLine("DEBUG: " + data.toString());
+        });
+        // TODO: Utilize streaming api to iniate schema reloads
+        dataConnectEmulatorDetails.instance.stderr?.on("data", (data) => {
+          if (data.toString().includes("Finished reloading")) {
+            this.schemaReload();
+          } else {
+            emulatorOutputChannel.appendLine("ERROR: " + data.toString());
+          }
+        });
+      }
+    }
   }
 
   dispose() {