+++ src/commands/dataconnect-sdk-generate.ts
@@ -1,4 +1,5 @@
 import * as path from "path";
+import * as clc from "colorette";
 
 import { Command } from "../command";
 import { Options } from "../options";
@@ -20,11 +21,27 @@ export const command = new Command("dataconnect:sdk:generate")
         const cwd = options.cwd || process.cwd();
         configDir = path.resolve(path.join(cwd), configDir);
       }
-      const serviceInfo = await load(projectId, service.location, configDir);
+      const serviceInfo = await load(projectId, configDir);
+      const hasGeneratables = serviceInfo.connectorInfo.some((c) => {
+        return (
+          c.connectorYaml.generate?.javascriptSdk ||
+          c.connectorYaml.generate?.kotlinSdk ||
+          c.connectorYaml.generate?.swiftSdk
+        );
+      });
+      if (!hasGeneratables) {
+        logger.warn("No generated SDKs have been declared in connector.yaml files.");
+        logger.warn(
+          `Run ${clc.bold("firebase init dataconnect:sdk")} to configure a generated SDK.`,
+        );
+        logger.warn(
+          `See https://firebase.google.com/docs/data-connect/gp/web-sdk for more details of how to configure generated SDKs.`,
+        );
+        return;
+      }
       for (const conn of serviceInfo.connectorInfo) {
         const output = await DataConnectEmulator.generate({
           configDir,
-          locationId: service.location,
           connectorId: conn.connectorYaml.connectorId,
         });
         logger.info(output);